refactor(e2e tests): replace relative import paths with aliases and move files to correct folders
