+++ tests/e2e/flowPages/fluggastrechte/formular/startFluggastrechteFormular.ts
@@ -1,7 +1,7 @@
 import { Page, expect } from "@playwright/test";
+import type { FluggastrechteFormular } from "tests/e2e/pom/FluggastrechteFormular";
+import { expectPageToBeAccessible } from "tests/e2e/util/expectPageToBeAccessible";
 import { today, toGermanDateFormat } from "~/util/date";
-import type { FluggastrechteFormular } from "../../../pom/FluggastrechteFormular";
-import { expectPageToBeAccessible } from "../../../util/expectPageToBeAccessible";
 
 export async function startFluggastrechteFormular(
   page: Page,
