+++ tests/e2e/flowPages/shared/finanzielleAngaben/finanzielleAngabenKinder.ts
@@ -1,6 +1,6 @@
 import type { Page } from "@playwright/test";
 import { Formular } from "tests/e2e/pom/Formular";
-import { expectPageToBeAccessible } from "../../util/expectPageToBeAccessible";
+import { expectPageToBeAccessible } from "tests/e2e/util/expectPageToBeAccessible";
 
 export async function startFinanzielleAngabenKinder(
   page: Page,
