+++ tests/e2e/pages/beratungshilfe.spec.ts
@@ -1,5 +1,5 @@
 import { test } from "@playwright/test";
-import { testPageToBeAccessible } from "../../util/testPageToBeAccessible";
+import { testPageToBeAccessible } from "../util/testPageToBeAccessible";
 
 test.beforeEach(async ({ page }) => {
   await page.goto("/beratungshilfe");
